fix(forge): address remaining reviewer findings on hook system PR
- Replace /dev/urandom session ID with uuid::Uuid::new_v4() (simpler, portable, no silent zero-fill fallback)
- Add debug! log when stdin write_all fails in CommandExecutor (was silently discarded)
- Fix rate limit constant type u32->u64 and update cross-reference comment to daemon.ts:1755
- Document POST /api/hooks/prompt-eval and /api/hooks/agent-eval in docs/API.md

+++ packages/daemon-rs/crates/signet-daemon/src/main.rs
@@ -184,10 +184,10 @@ async fn main() -> anyhow::Result<()> {
     // the two buckets don't share state and operators can tune them independently.
     let recall_llm_limiter = AuthRateLimiter::from_rules(&merged);
     // Dedicated limiter for hook prompt/agent-eval endpoints.
-    // Mirrors TS daemon's authHookEvalLimiter(60_000, 30).
-    // Constants must stay in sync with daemon.ts — update both sides together.
+    // Mirrors TS daemon: authHookEvalLimiter at daemon.ts:1755.
+    // Keep these values in sync when changing either side.
     const HOOK_EVAL_WINDOW_MS: u64 = 60_000;
-    const HOOK_EVAL_MAX: u32 = 30;
+    const HOOK_EVAL_MAX: u64 = 30;
     let hook_eval_limiter = {
         let mut rules = std::collections::HashMap::new();
         rules.insert("hook-eval".to_string(), RateLimitRule { window_ms: HOOK_EVAL_WINDOW_MS, max: HOOK_EVAL_MAX });
